python-tools/requirements.txt
google-cloud-bigquery
pyOpenSSL
google-cloud-dataplex
google-cloud-datacatalog-lineage
google-cloud-dataform
